Edited test scripts with proper cd commands
Added the cd command back into test-scripts when dropping the collections. This was missing before causing the drop collections to function properly as we had to cd into the main folder to run the data-differ.py file but there was no code to get back into test-scripts folder to drop the collections.